Work around fill_input_buffer issues by just setting to NULL
We don't know if the libjpeg config has specified fill_input_buffer
as jpeg_boolean, boolean, int, or some other thing. C++ is picky
about this.

Rather than try to figure out what it should be, just set the pointer
to NULL since it really is an assertion failure if it's called anyway.