Disable quickhash-1 for TIFFs with smallest layer > 5 MB
The quickhash-1 for TIFF files includes the smallest (top) layer of the
image.  For non-pyramidal generic TIFFs (or pathological pyramidal TIFFs)
this could be hundreds of megabytes, causing quickhash calculation in
openslide_open() to take an arbitrary amount of time.

It's not clear how to subset the tile data for these files such that the
quickhash is reasonably collision-resistant.  In the meantime, this patch
fixes the performance problem.

Workaround for #79.