mirax: Avoid subtiling if there are no overlaps or position map
Slides exported from the Pannoramic viewer have no overlaps or position
map, but we are still needlessly processing subtiles for the higher
layers.  Worse, we assume CameraImageDivisionsPerSide == 1 for these
slides, so we are processing 16x as many subtiles as we would on the
original slide.

For such slides, do not impose a lower bound on the number of tiles per
layer, and do not subtile.  When checking for active camera positions,
we now need to accept a tile when any of its constituent positions are
active.

On my system, this patch reduces the time to read layer 8 on an exported
version of CMU-1 from 250 ms to 10 ms.