mirax: Support downsampled slides
The IMAGE_CONCAT_FACTOR for a zoom level gives the exponent on the
layer's downsample, relative to the previous layer.  OpenSlide requires
that zoom level 0 have IMAGE_CONCAT_FACTOR == 0, and all other levels
have IMAGE_CONCAT_FACTOR == 1.

Downsampled slides saved by the Pannoramic viewer have a non-zero
IMAGE_CONCAT_FACTOR on level 0 and omit the corresponding number of zoom
levels.  (For example, an image saved 1:16 has an IMAGE_CONCAT_FACTOR of
4 and omits 4 levels.)  All other metadata remains the same as in the
original slide, except for the slide position file, in which all
coordinates are divided by the downsample factor.  (Actually, they're
within a couple of pixels of that value, for unknown reasons.)

Support arbitrary IMAGE_CONCAT_FACTORs on each zoom level.  Tested with
downsampled slides and with a slide hacked to omit an intermediate zoom
level.

Closes #85.