cache: Fix potential use-after-free
The cache does not guarantee that a value returned by
_openslide_cache_get() will be kept in memory until the caller is
finished with it.  The value is moved to the front of the MRU list, and
existing locking in TIFF and JPEG currently prevents races, so in
practice a use-after-free can only occur with NGR and under high
concurrency.  Nevertheless, have _openslide_cache_get() and _put()
acquire a value reference for the caller which must be released with
the new _openslide_cache_entry_unref().

Closes #89.