Drop CMake build system
We currently have two parallel build systems, Autotools and CMake.
The CMake scripts can produce Linux binaries but are really only intended
for MSVC builds, which we do not intend to support any longer.  The CMake
scripts are incomplete, impossible to keep aligned with the Autotools
ones, and an attractive nuisance for Linux users who don't know about
their deficiencies.

We are dropping support for MSVC because it's a C++ compiler that doesn't
support C99 syntax, which makes it problematic for building a library
in modern C.  MinGW-w64 is a workable alternative.