Stop reporting I/O errors through glib logging
The glib logging system is intended for reporting programming errors,
not runtime errors.

There are two types of openslide_open() failures:

1. We don't support this slide format at all.  openslide_open() returns
NULL.

2. We support this file but failed to open it.  openslide_open() currently
returns NULL and one or more messages is logged through glib.  The
application must register a glib log handler if it wants to catch these
messages.  This is the only place in the OpenSlide API where we expose
our use of glib.

We change case 2 to have openslide_open() return an openslide_t already
in error state.  This is consistent with the documented rules for error
handling.  We do not continue trying other formats.

We also change _openslide_set_error() not to log to glib when an
openslide_t transitions to error state.