Check unreliable_mcu_starts when searching backwards
_compute_mcu_start() was not checking unreliable_mcu_starts when
searching backwards for a valid start.  So, rather than searching only
until the beginning of a row in Hamamatsu VMS, we were continuing until a
previously-computed start or the beginning of the file.  An early read
of a down-file tile would cause us to compute thousands of MCU starts in
one go.