Don't require stdbool.h in openslide.h
fc61589 was overzealous about reintroducing stdbool.h: we can assume it
exists when compiling OpenSlide, but not when compiling programs that use
it, since those may build with MSVC.

The logic removed in that commit was incorrect, so we don't reintroduce
it here.  (It assumed HAVE_STDBOOL_H was available from someone's
config.h, and also assumed bool is 4 bytes instead of the more likely 1.)
Instead, assume for now that MSVC is the only compiler without C99
support.

Closes #115.