mirax: Round up tile counts on non-subtiled slides
When computing tiles per side on a subtiled slide level, we divide by
at most image_divisions, which always divides evenly.  But when
processing a non-subtiled slide (no overlaps and no position map), we
divide by arbitrary powers of two and were forgetting to round up.  This
could produce collisions in the JPEG tiles table, causing tiles to be
rendered in the wrong places.

Fixes #119.