hamamatsu: Stop sorting all_jpegs
It's a short list of large files, so we're only saving a few seeks if
we're saving any.  We're probably not saving any, because the on-disk
order may not match the lexical order of the filenames.  The only benefit
to sorting is to ensure that we scan the JPEGs in a consistent order.
GKeyFile is order-preserving, so we don't need the sort; we'll just use
the order in which the files are listed in the .vms.