cache: Change level argument to void *
There might be more than one coordinate plane per level, so we shouldn't
receive a level pointer.  For a while we were taking a grid pointer, but
the grid is an optional helper.  And we don't dereference the pointer
anyway; it's just a cookie for a coordinate plane.  Switch to taking a
void * selected by the vendor backend.