Reject overlarge key files without first reading 1 MB
The hamamatsu driver checks for VMS/VMU by trying to read the slide as
a key file.  We refuse to read more than 1 MB, but we weren't bailing out
until after reading it.  As a result, we were always reading the first
1 MB of e.g. TIFF files into RAM.  Instead, check the file size before
doing any reading.

This also allows us to allocate the buffer in one pass rather than
repeatedly reallocating.