tifflike: Read non-inline TIFF values on demand
tifflike is currently used only by NDPI, but we probe for NDPI before any
of the other TIFF-derived formats.  So, for non-NDPI TIFFs, we were
reading in the entire TIFF directory structure and then throwing it away.
Reduce I/O overhead by reading only the directories and inline values,
and then loading non-inline values on demand.