Pass tifflike, not tiffcache and TIFF, to TIFF format methods
We have to create a tifflike anyway for NDPI, so for efficiency we might
as well use it for all format detections.  Also pass it to the open
methods for their convenience, but require them to go get a TIFF * if
they want one.

Also: when we were passing a TIFF * we couldn't try opening the TIFF file
until after the NDPI detector ran, since otherwise libtiff would spew
error messages on NDPI slides.  We will now be able to try opening the
TIFF before running any format detections.

Always destroy the tifflike after the open method completes, even on
success.  If some backend needs otherwise, tifflike will need to grow
ref and unref methods.