Add openslide_detect_vendor() API function
Pursuant to #126.  The documented semantics of openslide_can_open()
prevent making it much faster than openslide_open(), so we need a
lighter-weight alternative.  As a bonus, have it return the vendor name,
so that VIPS doesn't have to call openslide_open() to detect generic-tiff
slides it doesn't want to open with OpenSlide in the first place.