openslide.pc: Drop static linking support
If we list a dependency in Requires.private, pkg-config requires its
.pc file to be installed, even when linking dynamically:

    https://bugs.freedesktop.org/show_bug.cgi?id=4738

This causes our -devel binary package to depend on many other -devel
packages which would not otherwise be needed.  (In fact, since Linux
distributions don't and won't ship static versions of OpenSlide, they
will never be needed -- but always required.)

Or it *would* cause this, except that Requires.private is woefully out of
date.  Fixing it is not completely trivial, since zlib, libtiff, and
OpenJPEG may be detected either by pkg-config or by hand, depending on
the versions of those libraries.  We would need to dynamically generate
Requires.private and Libs.private based on which detection method works
in each case.

I doubt anyone is linking OpenSlide statically and I also doubt that they
should.  If they are, they're probably building the entire dependency
stack themselves, so if we add a dependency, updating the linker command
line is going to be the least of their problems.  So, since the use case
is probably mythical, we're not going to carry configure.ac code to
support it.  Instead, we will drop Requires.private and Libs.private.

Note that you can still statically link against OpenSlide (why??), you
just have to spell out all of its dependencies on *your* linker command
line.