hamamatsu: Use tag 65420, not Software field, to detect NDPI
jcupitt/libvips#127 reports an NDPI file with a Software tag starting
with "NDP.toolkit" instead of "NDP.scan".  We could check for "NDP." but
that seems dubious.

TIFF tag 65420 seems to be present, with the value 1, in every NDPI file.
Since that is the first tag in the block of NDPI custom tags, assume it
is a marker for NDPI files, or perhaps a file format version.