sakura: Switch to simple grid
We're not going to be able to avoid parsing/generating tile ids, so
we might as well switch to the simple grid for slightly better
performance.  This should also allow us to pick up tiles that are added
to the database after the slide is opened.

No longer check for tiles with missing color components during open
since it is no longer convenient; just ignore these tiles at runtime.
That is probably not a bad idea anyway, since we may be racing with a
writer.