Don't use locals after longjmp that are modified after setjmp
Their values are indeterminate.  We got away with this for a long time,
but it was causing optimized builds on Mac OS 10.9 to crash in JPEG
error paths.

Fix by allocating state structs with gslice before calling setjmp.
This requires us to redo JPEG error handling, but the resulting API
is more encapsulated and more explicit.