Add macro to disable gcc attributes in public headers
MATLAB parses C headers with a Perl script that is confused by gcc
attributes.  If OPENSLIDE_SIMPLIFY_HEADERS is defined, avoid attaching
gcc attributes to function declarations.

In order to actually use this feature, MATLAB bindings will need to
create a wrapper header like this:

#define OPENSLIDE_SIMPLIFY_HEADERS
#include <openslide.h>

because MATLAB doesn't directly provide a way to define a macro before
parsing a header.