philips: Calculate level downsamples from pixel spacing attributes
The level widths/heights given in the TIFF directories and in the
ImageDescription XML include the dimensions of the padding in the
rightmost/bottommost tiles, including for level 0.  Since the same
tile size is used for every level, each level has a different amount
of padding and a different aspect ratio.  This causes image shifts
between levels.

We can't directly calculate the correct dimensions for a level, and
we can't easily guess by looking for white boundary pixels because
slides with multiple ROIs may not have a consistent right/bottom margin.
However, the metadata includes the pixel size for each level (in mm), and
we can divide this by the level 0 pixel size to determine the level
downsample.  Set the level downsamples in this way so viewers can
avoid level shifts.