hamamatsu: Don't try to read optimisations for VMS map file
None of my sample VMS files have optimisation data for the map file.
On slides with a single Z-layer this would cause us to harmlessly give
up on the .opt file, but slides with multiple Z-layers place the
additional optimisation data after the optimisations for Z-layer 0.
This was causing restart marker position mismatches for those slides.

Stop trying to read optimisations for the map file.

Fixes #165.