test: Use platform-specified stack frame alignment on Win32
The Win32 calling convention only requires 4-byte alignment of stack
frames, but GCC expects 16-byte alignment.  This can cause crashes when
interworking with MSVC, such as openslide/openslide-winbuild#9, if
OpenSlide is not built with compensatory compiler flags (such as
-mstackrealign).  On Win32 only, build the test programs with the stack
frame alignment specified by the Win32 calling convention, in order to
check that interwork is functional.  This can be disabled by configuring
with --disable-checking-windows-abi-compat in case interwork is not
needed.