Add exported cache management API
Create a slide handle with a default cache, but allow the user to
separately allocate a cache of any size and attach it to one or more
slide handles.

For now, don't support changing the size of an existing cache, nor
external caches implemented by the application.  Those can be added
within the existing API if needed.