cache: add a serial number for the binding to the cache key
When attaching a cache binding to a cache, assign it a serial number
from a counter maintained by the cache, and then include that serial
number in every cache key from that binding.  This ensures that stale
cache values will not be returned if the slab allocator reuses an old
level struct in a subsequent slide handle after the original slide handle
has been closed.