Add vendor driver with synthetic slide, gated behind debug flag
Currently the only way to test OpenSlide's use of image decoding
libraries is to try reading a series of slide files that use the
corresponding image formats.  This makes it non-trivial to smoke-test
a build.  Add a vendor driver that renders a synthetic slide whose
individual tiles are small compiled-in images of each supported type.
To avoid interfering with opening legitimate slides, return this slide
only when the empty path ("") is opened and OPENSLIDE_DEBUG includes the
new "synthetic" flag.

Try decoding each tile during openslide_open() so a smoke test doesn't
need to call openslide_read_region().  When decoding a tile, check some
representative pixels to verify that they were decoded correctly.

Add properties describing the test images that were rendered and giving
their indices for debugging purposes.  The tile size can be determined
via the usual tile size properties.

Put the new driver at the top of the search order so other drivers
don't check for weird invalid filenames at the root of the filesystem.

As an experiment, use g_autoptr for cleanup where possible.  The API
isn't a good fit for slice-allocated buffers, and insists on typedefs
for struct types (which I've used here as minimally as possible), but
generally seems to be an improvement over manual cleanup.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>