test/driver: unpack xdelta onto reflinked base file, if possible
If the underlying filesystem supports it (XFS or btrfs), use reflinks
to reuse unmodified data blocks from the pristine copy of the file.  We
could do this at blksize granularity (4 KiB), but save some syscalls by
doing it in 1 MiB chunks.  In practice, most of our xdeltas only apply a
handful of distinct modifications to a file; we're mainly trying to avoid
copying hundreds of MiB of bulk data.

This reduces _slidedata from 124 GiB to 22 GiB on XFS.

Closes https://github.com/openslide/openslide/issues/309.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>