workflows: add Debian stable, Ubuntu 20.04/22.04, CentOS Stream 8/9
Test on a variety of distros that we want to keep working.

Notably, we're not testing Debian oldstable, Ubuntu 18.04, and CentOS 7,
which in practice means that we're dropping support for them.  The current
Debian oldstable will go EOL next month.  Ubuntu 18.04 has not been the
current LTS release for two years, and likewise RHEL 8 was released in
2019.  While it's good to preserve some compatibility with LTS distros,
we shouldn't maintain very old compatibility constraints forever, and
this seems like a good compromise.

Certain test steps only need to be run once; use the Fedora container.

Ignore file descriptor leaks in distro versions that have leaky standard
libraries.

On CentOS Stream 9, install xdelta from a Copr pending
https://bugzilla.redhat.com/2110358.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>