Don't compile with -pedantic
We're no longer trying to strictly comply with C99, since we're now
dependent on the GCC/clang cleanup attribute.  Regardless, the GCC docs
say -pedantic doesn't issue diagnostics for all deviations from the C
standard, so we can't assume that use of the option enforces compliance.
We already enable -Wall and -Wextra, so let's stick to those, rather than
asking for warnings the compiler thinks are dubious.

This makes -Wno-pedantic-ms-format unnecessary, so drop it.

This was originally motivated by openslide-winbuild's use of #include_next
in https://github.com/openslide/openslide-winbuild/pull/49, coupled with
the use of -Werror in CI.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>