xml: add a wrapper for xmlFree()
xmlFree() is a macro that calls through a pointer to a function whose
declaration is invisible to CFI, so CFI will warn about it.  Centralize
calls in a wrapper function that we can then ignore.

We can't effectively use _OPENSLIDE_POISON on xmlFree because the libxml
headers unconditionally #undef xmlFree before redefining it.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>