Fail openslide_open() on unpatched pixman 0.38.x
pixman 0.38.x produces corrupt output when subpixel rendering is used
with the SATURATE operator.  This is fixed in 0.40.0:

    https://gitlab.freedesktop.org/pixman/pixman/-/commit/8256c235

The consequence is pretty bad (blank output in some regions of some
slides) so detect it and refuse to open any slide.  Perform a behavior
test at runtime, since we might have been compiled with a different
version or the distro might have backported a fix.

See: https://github.com/openslide/openslide/issues/278
Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>