update comment and remove stray blank line
In the ImageDescription tag embeded in SCN file, all image XML nodes
look similar. Without SCN specification, one can only guess the image
taken with lowest power objective is the macro image. So far this
assumption agrees with the Aperio ImageScope viewer. It may also be
image of the first XML node in ImageDescription tag, but there is no way
to confirm.

Signed-off-by: Wei Chen <chenw1@uthscsa.edu>