Use _WIN32 and __APPLE__ to gate OS-specific code
We currently have three ways to check for Windows: WIN32, _WIN32, and
G_OS_WIN32.  Consistently use _WIN32, which is defined by the compiler.
WIN32 isn't safe, since it's defined somewhere in the platform;
apparently Visual Studio generates projects that define it in the
project configuration by default.  (MinGW does define it, but evidently
not when compiling with Meson.)  G_OS_WIN32 requires glib headers, which
notably are not available in our public headers.

In addition, replace some feature checks with the OS checks they're
actually a proxy for.  fcntl() should always exist on non-Windows,
_wfopen() will always exist on Windows, and proc_pidfdinfo() will
always exist on Darwin.  If we actually end up needing a feature check
somewhere, we can add it at that time.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>