Drop Cygwin cross-test support
It doesn't make sense to test from Cygwin anymore; WSL or native builds
are the way forward.  Also, current Python has native Windows symlink
support.

We'll need to figure something else out, but in the meantime, this code
hasn't been used in a long time and it's in the way.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>