test/driver: set execute bit in source tree
Meson configure_file() doesn't provide a way to set the execute bit in the
build directory, only after installing, and we don't install the test
driver.  We could add a custom target to chmod the file, but that seems
awkward and brittle.  Instead, just set the executable bit on the source
file, and add a check that fails if the source file is executed directly.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>