meson: add `doc/html` build target for Doxygen documentation
Don't build docs by default.  The Autotools build system doesn't, and
since Doxygen tends to change the set of valid Doxyfile directives on
every patch release, running Doxygen on an older or newer Linux distro
produces a pile of warnings.

Compensate for this by continuing to ship the generated docs in the
release tarball.  But, when we're built as a subproject and the parent
project is being distributed via meson dist --include-subprojects, skip
docs generation during dist.

Co-authored-by: Jan Harkes <jaharkes@cs.cmu.edu>
Signed-off-by: Jan Harkes <jaharkes@cs.cmu.edu>
Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>