tiff: hold file handle open until cached_tiff is released
Opening and closing an _openslide_file for every TIFF read is needlessly
inefficient.  Once the file is opened, hold it open until the
_openslide_cached_tiff is released, since that object is single-threaded
and will be released before OpenSlide returns to its caller.  Don't
eagerly open the file when the _openslide_cached_tiff is acquired, since
the needed pixels may be in cache and we may not need to read from the
TIFF at all.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>