meson: add optional libdicom dependency
libdicom isn't widely available yet, so make the dependency optional.  To
make DICOM support easier to use, also include a libdicom wrap for now.
Eventually we may want to drop that and require libdicom to come from the
system as we do for other dependencies.  But at the moment, nothing else
uses it and we're dependent on a specific commit, so shipping a wrap is
useful and safe.

Also add a uthash wrap, since Meson < 0.56 doesn't automatically promote
libdicom's wrap to the top-level project.  It's a header-only library, so
we're not risking installing conflicting shared objects.

Disable libdicom's tests to avoid pulling in 'check' as a dependency.

Co-authored-by: John Cupitt <jcupitt@gmail.com>
Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>
Signed-off-by: John Cupitt <jcupitt@gmail.com>