Add initial DICOM vendor driver using libdicom
It can correctly load JPEG TILED_FULL DICOM images and label/macro
associated images.  The 3DHISTECH demo DICOMs work well.

We can't assume particular filenames for the image files, so the driver
takes the filename of one image and searches the enclosing directory
for other images with the same slide ID.

Missing features:

- Sparse images
- Properties
- quickhash-1
- Tile formats other than JPEG
- ICC support
- Closing file handles between API calls

There are a lot of debug prints still in the code, but compiled out.
They'll be removed once the driver stabilizes a bit.

Co-authored-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>
Signed-off-by: John Cupitt <jcupitt@gmail.com>
Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>