Add API for reading associated image ICC profiles
There's no requirement that a format uses the same ICC profile for the
main image and all associated images.  Add separate APIs for getting
associated image ICC profiles, to at least prevent applications from
assuming that the main ICC profile can be used everywhere.

When the profile for an associated image is known, drivers should return
it, even if it's the same as for the main image.  Do this for DICOM
slides, which generally provide the ICC profile in each individual file.
For other formats where we currently support profiles, it's not obvious
whether the main profile should also apply to associated images, so return
nothing for now.

Fixes https://github.com/openslide/openslide/issues/139.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>