dicom: support quickhash-1
The purpose of the quickhash is to "uniquely identify a particular
virtual slide, but [not] to detect file corruption or modification".
Most slide formats don't include unique identifiers, so we usually have
to construct one from metadata that collectively is unique to a slide.

In the case of DICOM, there are already unique identifiers for entities
at various conceptual levels.  Those identifiers refer to the
conceptual entity, not its exact representation.  For example, the SOP
Instance UID doesn't change if the pixel data or metadata are altered
in certain ways that don't change the instance's notional semantics.

Rather than second-guessing DICOM's approach to entity naming, adopt it.
That requires us to pick a UID to hash.  The Acquisition UID is too
general (two recompressions of the same scan share the same UID) and its
presence is optional; the Pyramid UID is optional; and the SOP Instance
UID refers to a specific level, portion of a level, or associated image.
We could hash all the SOP Instance UIDs, but then we'd be creating our
own definition of a "slide" out of its constituent parts.

The Series Instance UID, however, seems like a good fit, especially since
we currently support only one series per slide.  Calculate the quickhash
by hashing that UID.  In the future, if we support Z-stacks distributed
across multiple series (with a common Acquisition ID), we can hash a
sorted list of all Series Instance UIDs.

This will cause quickhash collisions if a single series (or a reused
series UID) contains multiple independent slides _and_ we aren't
presented with the entire series at once.  We could try to avoid that by
hashing additional attributes such as timestamps, but again we'd be
second-guessing the existing naming semantics.

Discussion in https://github.com/openslide/openslide/pull/457.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>