meson: drop fopen CLOEXEC test program; just look at host OS
The CLOEXEC check is our only use of compiler.run(), which breaks
cross-compilation when the build machine can't run the host's binaries
(except on Windows, which we special-cased).  In fact, all modern versions
of Linux (glibc and musl) and of the BSDs support the "e" flag.  Drop the
test program and just look at the host OS string.

Reported by @alanocallaghan in openslide/openslide-winbuild#96.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>