Require libdicom
Now that the libdicom API has stabilized and DICOM support is no longer
under rapid development, require libdicom, so every OpenSlide
installation will support the same formats.  For now, we'll continue
shipping the Meson wrap as a convenience; we can drop it once libdicom
is broadly available in package managers.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>