aperio: allow floating-point objective power
Some Aperio slides have a floating-point AppMag such as "40.000000".
In the samples I've seen, the ImageDescription mentions both "Aperio
Image Library" and "Leica SCN".  One such file:

    TCGA-3C-AALI-01Z-00-DX2.CF4496E0-AB52-4F3E-BDF5-C34833B91B7C.svs
    https://portal.gdc.cancer.gov/files/6a7477b1-86f3-473f-9bf1-174a758142e9

Support transcribing such AppMags to openslide.objective-power.  The
value will still be stored as an integer if it can be represented in
one.

Fixes: https://github.com/openslide/openslide/issues/247
Fixes: https://github.com/openslide/openslide/issues/332
Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>