meson: add a private option to filter out non-OpenSlide symbols
On macOS, openslide-bin needs to use a linker option to filter out
non-OpenSlide exported symbols.  It can't set the option globally because
that would break OpenSlide Java, and can't set a per-project option
because Meson doesn't support that.  Add a private option to do the
filtering until Meson provides syntax for this.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>