tools: suppress libpng warning about edited sRGB profiles
When writing a PNG, libpng recognizes specific sRGB ICC profiles and
automatically adds an sRGB chunk.  Sometimes, e.g. with
DICOM/CMU-1-JP2K-33005, it detects profiles which are sRGB but don't
exactly match what it's looking for, in which case it logs a warning and
doesn't add the sRGB chunk:

    libpng warning: Not recognizing known sRGB profile that has been edited

We're just copying a profile from the original slide to the output PNG, so
this warning isn't actionable.  Suppress it.  There's no machine-readable
error code, but the error message isn't localized and hasn't changed since
at least 2014.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>