zeiss: stop validating values against file size
It isn't a complete solution to any problem: the file could have been
truncated after we opened it, so we still need to care about short reads,
and the file could be larger than memory, so we still need to care about
allocation failures.  We already check for short reads, and can avoid
crashing on large allocations by using g_try_*().  Once we have those
mitigations, we don't need the file size check, so drop it.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>