zeiss: drop czi_decbuf
It's an in-out structure, which adds complexity, and the struct as a whole
doesn't really add value.  It also requires an extra memcpy when reading
associated images.

Without czi_decbuf, czi_bgrn_to_argb32() doesn't need to be a separate
function, so consolidate it into czi_read_uncompressed().

Clean up data types and cache key calculation in read_tile().

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>