zeiss: drop czi_scene; move common_downsample out of struct czi
struct czi_scene can be replaced with local arrays if we combine scene
evaluation and setting of region properties into a single function.

That function can also compute the common max downsample.  That value is
needed in two places, create_levels() and init_range_grids(); the latter
can derive it from the level array, and the former can receive it as an
argument.  This seems clearer than deriving it as a side effect of
create_levels() and storing it in a long-lived struct even though it's
only going to be used one more time.

Also add bounds checking on czi_subblk.scene and ensure every scene has
at least one subblock.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>