zeiss: fix tile size declarations
We were setting the tile-width and tile-height properties for each level,
but those hints should only be present when tiles are arranged in a
regular grid.  Also we were hardcoding them to 256 pixels.  Remove the
properties.

In addition, the range grid wants to know the typical tile size so it can
size its buckets correctly.  We were passing 256x256, which is
substantially too small.  Pass the largest tile width and height for each
level, on the belief that large tiles are normal and smaller tiles are at
boundaries.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>