workflows: automatically create draft GitHub release when tagged
Run `meson dist` alongside the extra checks in every build, because it's a
good idea to test this and because the build job already installs the
dependencies for `meson setup`.  Archive the artifact, then conditionally
run a second job to create the release.  Leave the release in draft state
so a longer-form release description can be added before publishing.

This uses the same release-job pattern as OpenSlide Python, rather than
starting a second workflow upon completion of the build workflow.  This
results in a skipped job in every PR's job list, but allows the passing of
output variables between jobs.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>