zeiss: parse metadata XML values from vendor properties
Read standard properties and metadata values from vendor properties.  This
is documented to be the preferred approach, since it ensures the
underlying raw data is available from properties.

Don't assume that the first objective in the Objectives list is the
relevant one; dereference Information.Image.ObjectiveSettings.ObjectiveRef
instead.

Drop the XML XPath helpers we added.  We don't use them anymore, and
they're probably only useful for parsing values directly out of XML
metadata, which we want to discourage.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>