workflows: run Fedora job before other build jobs
The Fedora job runs a number of extra tests.  By declaring it separately,
we duplicate some code (mainly dependency installation) but we also drop a
number of conditionals.  More importantly, by sequencing the other jobs
after the Fedora one, we can avoid building the pristine cache multiple
times in parallel when missing.

Fail the subsequent jobs if the cache can't be restored.  Install the zstd
command-line tool to avoid an obscure restore failure in actions/cache.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>