Implement our own BMP loader; drop gdk-pixbuf dependency
gdk-pixbuf is pivoting to become a wrapper around a sandboxed Rust library
that reads each image in a separate process and currently only works on
Linux.  Upstream also seems restless to abandon the existing loaders,
whether or not replacement functionality is in place.  The loader
mechanism will still exist, but out-of-tree loaders can't be built
directly into gdk-pixbuf, so openslide-bin would need to patch it to keep
BMP compiled in.  It's time to migrate away from gdk-pixbuf.

The available widely-used BMP decoders in C tend to be multi-format
libraries that bring in a lot of other dependencies.  We used gdk-pixbuf
to avoid reimplementing BMP support, but lacking plausible alternatives we
can just handle it ourselves.  BMP is not especially complex, particularly
for the uncompressed MIRAX data we need it for.

Please report any MIRAX BMP regressions this causes.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>