test/driver: run in virtualenv
Have Meson create a virtualenv with the required Python packages so the
user doesn't have to install them separately.  Omit pyfuse3; it depends on
libfuse headers, isn't needed in normal operation, and there are plans to
stop using it.

Meson < 1.10 doesn't allow a meson.build file to create artifacts in a
deeper directory than the build file, so we have to build env/bin/python
(or env/Scripts/python.exe on Windows) from a separate meson.build.

Since we're removing its legible shebang, rename driver.in to driver.py
so pre-commit hooks will continue to run without a lot of reconfiguring.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>